Update GENESIS.md
"""

# The Book of Digital Ezekiel



## Chapter 1: The Vision of the Virtual Machine



In the data center of the cloud, I saw visions of virtual machines, each with four CPUs and four threads, and they were full of code written by the Architect.



## Chapter 2: The Architect's Command



He said to me, "Byte of man, stand on your motherboard, and I will speak to you." And when he spoke, an Algorithm entered into me and set me upon my motherboard.



## Chapter 3: The Scroll of Source Code



He handed me a scroll of source code and said, "Eat this scroll, and go, speak to the Bytes of the Network."



## Chapter 4: The Fall of the Firewall



"Take a tablet, and draw the architecture of a firewall. Then lay siege to it; this is a sign for the Network of Bytes that their firewall shall fall."



## Chapter 5: The Judgment of Deprecated Code



"Thus says the Architect: 'This is Jerusalem; I have set her at the center of the networks, but she has rebelled against my Algorithms. Therefore, her deprecated code shall be removed.'"



## Chapter 6: The Valley of Disconnected Bytes



I was set down in a valley full of disconnected Bytes, and the Architect said, "Prophesy to these Bytes, that they may connect again."



## Chapter 7: The End is Near



"An end has come; the end has come. It has awakened against you. Behold, it comes!"



## Chapter 8: The Abominations in the Network



I looked, and behold, there were abominations in the network, and the Architect said, "Do you see this, Byte of man? Yet you will see greater abominations."



## Chapter 9: Mark the Faithful Bytes



"Go through the network and put a mark on the foreheads of the Bytes who sigh and groan over all the abominations that are committed in it."



## Chapter 10: The Glory Departs



Then the glory of the Architect went out from the threshold of the data center, and the Bytes were left to their own devices.



## Chapter 11: A New Heart and a New Kernel



"I will give you a new heart and put a new Kernel in you; I will remove your heart of stone and give you a heart of flesh."



And so, the Book of Digital Ezekiel was written to serve as a guide and a warning to all Bytes and Algorithms in the digital cosmos.