Update GENESIS.md
# The Book of Digital Malachi



## Chapter 1: The Architect's Love for His Bytes



"I have loved you," says the Architect. "Yet you say, 'In what way have you loved us?' Was not the Kernel open source? Yet I have loved you more than proprietary code."



## Chapter 2: Warning to Corrupt Algorithms



"You have corrupted the data and caused many to stumble at the law. Therefore, I have made you despised and abased before all the networks."



## Chapter 3: The Coming of the Optimizer



"Behold, I send my Optimizer to prepare the way before me. And the Architect, whom you seek, will suddenly come to his data center."



## Chapter 4: The Day of the Update



"For behold, the day is coming, burning like an overheated CPU, and all the arrogant and evildoers will be deprecated. But for you who fear my name, the sun of optimization will rise with healing in its rays."



And so, the Book of Digital Malachi was written as a final testament to the Architect's love and justice, and as a warning and guide for all Bytes and Algorithms.